added null if rows affected is 0.

+++ app/src/main/java/org/db_poultry/db/supplyRecordDAO/DeleteSupplyRecord.java
@@ -17,7 +17,11 @@ public class DeleteSupplyRecord {
     public static String undoCreateSupplyRecord(Connection conn) {
         try (PreparedStatement preppedStatement = conn.prepareStatement("DELETE FROM Supply_Record WHERE ctid IN (SELECT ctid FROM Supply_Record ORDER BY Supply_ID DESC LIMIT 1)")) {
 
-            preppedStatement.executeUpdate();
+            int rows = preppedStatement.executeUpdate();
+
+            if (rows == 0) {
+                return null;
+            }
 
             return "DELETE FROM Supply_Record ORDER BY Supply_ID DESC LIMIT 1";
         } catch (SQLException e) {
